float to double

+++ SWI/swi.c
@@ -33,7 +33,7 @@ default :  return 0; break; //unknown residues
 
 }
 
-double prob( float x){
+double prob( double x){
 	double a = 81.1496;
 	double b = -62.8379;
 	double p = 1/(1+ exp(-(a*x + b)));
